Revert to using WhereFilter.call_parameter_sets.
